fix: 스크랩 페이지네이션 에러 수정
