+++ src/views/member/MyScrap.vue
@@ -9,10 +9,54 @@
       />
       
         <h2 style="text-align:center; margin: 20px;">📔 스크랩한 프로젝트</h2>
-        <ProjectScrapComponent
-        :projectList="this.projectList"
-        ref="myScrapComponent"
-        />
+        <v-container class="outer-box">
+
+          <v-container class="d-flex">
+              <v-card
+                class="mx-auto"
+                v-for="project in projectList"
+                :key="project.id"
+
+                >
+                <v-img
+                height="200"
+                :src="project.imageUrl"
+                @click="spaMoveTo(project.id)"
+                cover
+                class="scrap-img"
+                ></v-img>
+                <v-card-title>
+                    {{ project.projectName }}
+                </v-card-title>
+                <v-card-text>
+                     <p>{{ project.description.substr(0, 7) }} <span style="font-size:small; color: gray;">...more</span></p>
+                     <p style="font-size:small; color: gray;">모집 마감: </p>
+                     <p style="font-size:small; color: gray;">{{ this.getDay(project.deadline) }} {{ this.getTime(project.deadline) }}</p>
+                </v-card-text>
+      
+                <v-card-text  class="d-flex justify-space-between align-center">
+                  <BasicChip :title="this.getProjectStatus(project.isClosed)" :color="this.getChipColor(project.isClosed)"/>
+                  <v-icon style="margin-left: 15px" size="x-large" icon="mdi-bookmark" class="scrap-icon" @click="this.cancelBookmark(project.id)"></v-icon>
+                </v-card-text>
+            </v-card>
+          </v-container>
+      </v-container>
+      
+      <!-- 모달 -->
+      <v-dialog v-model="dialog" width="500px">
+      <v-card class="dialog-card">
+          <v-card-title>
+              북마크를 취소하시겠습니까?
+              </v-card-title>
+      
+              <v-card-actions>
+              <v-spacer></v-spacer>
+              <v-btn color="sid_btn1" text @click="dialog = false">닫기</v-btn>
+              <v-btn color="sid_btn2" text @click="confirmCancel()">북마크 취소</v-btn>
+              </v-card-actions>
+      </v-card>
+      
+      </v-dialog>
 
 
         <!-- 페이지네이션 -->
@@ -40,51 +84,123 @@
   <script>
   import axios from 'axios';
   import PageNavbar from '@/components/navbar/PageNavbar.vue';
-  import ProjectScrapComponent from '@/components/card/ProjectScrapComponent.vue';
+  // import ProjectScrapComponent from '@/components/card/ProjectScrapComponent.vue';
 
   export default {
     components: {
         PageNavbar,
-        ProjectScrapComponent,
+        // ProjectScrapComponent,
     },
     data() {
       return {
         projectList: [],
         currentPage: 0,
         totalPage: 0,
+        dialog: false,
       }
     },
     methods: {
       moveToEdit() {
             this.$router.push("/member/edit");
         },
-        async onPageChange() {
-          try {
+        spaMoveTo(projectId) {
+            this.$router.push('/project/' + projectId);
+        },
+        getChipColor(title) {
+            if(title === 'Y') {
+                return '#F15F5F';
+            } else {
+                return 'sid_btn2';
+            }
+        },
+        getJobColor(job) {
+            if(job === 'Backend') {
+                return 'be_blue';
+            } else if(job === 'Frontend') {
+                return 'fe_yellow';
+            } else if(job === 'PM') {
+                return 'pm_green';
+            } else if(job === 'AOS') {
+                return 'app_red';
+            } else if(job === 'iOS') {
+                return 'app_red';
+            } else if(job === 'Designer') {
+                return 'de_purple';
+            } else {
+                return 'white';
+            }
+        },
+        getProjectStatus(yn) {
+            if(yn == 'Y') {
+                return '마감';
+            } else {
+                return '진행 중';
+            }
+        },
+        cancelBookmark(id) {
+            this.dialog = true;
+            this.cancelId = id;
+        },
+        async confirmCancel() {
+            try {
+                await axios.post(`${process.env.VUE_APP_API_BASE_URL}/api/project/${this.cancelId}/scrap/delete`);
+            } catch(e) {
+                console.log(e);
+            }
+            this.dialog = false;
+            window.location.reload();
+        },
+        getDay(deadline) {
+            const createdTime = new Date(deadline);
+
+            return `${createdTime.getFullYear()}년 ${createdTime.getMonth()}월 ${createdTime.getDate()}일`; 
+        },
+        getTime(deadline) {
+            const createdTime = new Date(deadline);
+            let hour = createdTime.getHours();
+            let minute = createdTime.getMinutes();
+            let ampm;
+            if(hour < 12) {
+                ampm = '오전'
+            } else {
+                ampm = '오후'
+                hour -= 12;
+            }
+            if(hour < 10) {
+                hour = '0'+hour;
+            }
+
+            if(minute < 10) {
+                minute = '0'+minute;
+            }
+
+            return ampm + ' ' + hour + ':' + minute;
+        },
+    },
+    watch: {
+      async currentPage() {
+
+        try {
             const params = {
               size: 3,
               page: this.currentPage-1
             }
-            const response = await axios.get(`${process.env.VUE_APP_API_BASE_URL}/api/project/my-projects`, {params});
+            const response = await axios.get(`${process.env.VUE_APP_API_BASE_URL}/api/project/scrap`, {params});
             this.projectList = response.data.content;
             } catch(e) {
             console.log(e);
           }
-          this.$refs.myScrapComponent.changePage(this.currentPage);
-        },
-    },
-    watch: {
-      currentPage() {
-        this.onPageChange(this.currentPage);
       },
     },
     async created() {
       const params = {
-        size: 3
+        size: 3,
+        page: 0
       };
 
       const response = await axios.get(`${process.env.VUE_APP_API_BASE_URL}/api/project/scrap`, { params });
       this.projectList = response.data.content;
-      this.currentPage = 0;
+      this.currentPage = 1;
       this.totalPage = response.data.totalPages;
     },
 
@@ -92,21 +208,7 @@
   </script>
   
   <style scoped>
-  .my-page-card{
-    color: #E3E3E3;
-    width: 500px;
-    display: flex;
-    align-self: center;
-    align-content: center;
-    margin: auto;
-  }
-  .item-wrap {
-    display: flex;
-    justify-content: center;
-    padding-top: 20px;
-    padding-bottom: 20px;
-    margin: auto;
-  }
+
   /* 중앙 정렬을 위한 스타일 */
   .text-center {
     text-align: center;
@@ -115,45 +217,6 @@
     padding-top: 20px;
     font-weight: bold;
   }
-  
-  .text-left {
-    text-align: left;
-    width: 300px;
-    padding-bottom: 20px;
-    padding-top: 20px;
-  }
-
-  .bordered-table th,
-  .bordered-table td {
-    border-bottom: 1px solid #D4D4D4;
-    height: 60px;
-  }
-  
-  /* 열 너비 스타일 */
-  .column-width {
-    width: 1000px; /* 원하는 너비로 설정 */
-  }
-
-.my-div {
-    display: flex;
-    justify-content: center;
-}
-
-.card-url {
-    color:#D4D4D4;
-}
-
-.v-con {
-    display: flex;
-    align-self: center;
-    align-content: center;
-    text-align: center;
-}
-
-.row-center {
-    text-align: center;
-    align-items: centee;
-}
 
 .custom-container {
   max-width: 1200px !important; /* 원하는 최대 폭 */
@@ -163,5 +226,16 @@
 
 .self-center {
   margin: auto;
+}
+
+.outer-box {
+  display: flex;
+  width: 70%;
+}
+
+.scrap-icon:hover {
+  color: grey;
+  justify-self: flex-end;
+  
 }
 </style>
